Add: Better icons
I was constantly worried there is something wrong with my unit when seeing alert
icon in Lovelace :)

This commit change default icon to "fan", which is IMHO better for circulation
unit.

It also display different icons for different states. This is first try with
hacky implementation. Currently (first match wins):

- alert icon if there are some alerts
- fan with exclamation point for warnings
- striked when state is off
- arrows in circle for recirculation
- fan