Update climate.py
Changed the negative temperature computation according to the modbus documentation [ 65036 ~ -50,0°C ..65535 ~ -0,1°C, 1..1300 ~ 0,1..130,0°C ] as the original implementation has been throwing values arounf -5.900 C into HA.

Also changed register I10202 (Input TEa) to register I10211 (T-ODA) according to the Modbus parameters. T-ODA is always the temperature of the incoming outside air according to the setting in the service menu (it is generally TEa, but depends on the orientation of the unit and can also be set externally through modbus) - possibly not necessary.